requirements.txt
aws-cdk-lib==2.162.1
constructs>=10.0.0,<11.0.0
cdk-nag==2.32.0
cdk-ecr-deployment==3.0.135